+++ src/test/se/jbee/inject/bind/TestOptionBinds.java
@@ -9,8 +9,8 @@ import se.jbee.inject.Injector;
 import se.jbee.inject.bootstrap.Bootstrap;
 import se.jbee.inject.bootstrap.BootstrapperBundle;
 import se.jbee.inject.bootstrap.Bundle;
-import se.jbee.inject.bootstrap.OptionBootstrapperBundle;
 import se.jbee.inject.bootstrap.Module;
+import se.jbee.inject.bootstrap.OptionBootstrapperBundle;
 import se.jbee.inject.config.Globals;
 import se.jbee.inject.config.Options;